App.xaml.cs:    //* 20 line  commeted version of OnLaunched which demonstrates using a Button as the Window.Current.Content, instead of the usual Frame "rootFrame" object which can be used to Navigate backwards and forwards through Page History.
It is highly recommended to use a Frame as the Window.Current.Content because the Frame control is especially built for, and indeed can only contain Page objects.  With this in mind, the intent is to design UWP applications so that the Pages are central to the application, and the interface chiefly (if not soley) consists of Pages.

ControlCorral.csproj:
  This file is modified with the addition of 1 line
<Compile Include="GenericCommand.cs" />
to include  the new GenericCommand object, which is essential here to the implementation of  seperation of conerns between the GUI and the inner workings of the software itself.

MainPage.xaml
    Very humble beginnings of the "ControlCorral" to be developed in this Chapter:  Simply a single button centered inside of a Grid.

MainPage.xaml.cs:
  The meat and potatoes in the ControlCorrral application, which could possibly be more desciptively named "The Reservation Software", if it the Author's main intention was not to teach more about introducing more Controls and teaching one way of using them with a data store (a collection of ReservationInfo objects stored in RAM.)  I suspecte the Authors will eventually move on to data persistance issues, possibly by storing the data in files, bue I'm not sure.