Begin Chapter Two of "Real World Windows 10 Development (2nd Ed.) by Edward and Elizabeth Moemeka.
This Chapter expands on Chapter One, with some more Basic Controls, showing how to seperate the concerns of developing an Interface from the concerns of the application specifics.

New object "GenericCommand" will be used exensively for implementing this separation of concerns.

After attempting to learn Programming Windows 10 with various books I'd encountered--I must say that Edward and Elizabeth's book is where the rubber really meets the road for me.  I love this book, and their practical approach to teaching Windows 10 is, in my opinion Logical and Fun at the same time.  I am very pleased to have this book in hand.