After much struggling with getting all the details in the right places, regarding the hows of setting up Command Binding, I finally am beginning to cipher what's going on here.
I won't be able to do it without referring to the docs yet, but prior to this I've been struggling with the docs, so I'm counting this as progress.

I'll now make a new "CommandBinding" branch, which logically could have been made just prior to this development, but I'm just learning git too, and so I'm satisfied taking baby steps.

--amendment: I wanted to mention that watching MaingPage.g.cs (the source generated by vs, to merge the partial class with the interface.)  was integral in learning more about command binding.  I think it will very instructive to watch the object.g.cs files moving forward.