Following the book, refactoring the code to use a "ControlCorralModel".
I'm not completely convinced that this example will be practical in
improving the example, but I'll attempt to keep following along because
I'm sure to learn a lot on the way.  For example, the author(s) are
using some Linq to generate the data for the Model, and I'm very
unfamiliar iwht linq--so just typing the examples in verbatim is
educational for me.